refactor : 자료형 변경

+++ src/main/java/kr/or/ddit/highSchool/service/HighSchoolService.java
@@ -8,7 +8,7 @@ public interface HighSchoolService {
 	public List<HighSchoolVO> getAllHighSchools();
 
 	// 고등학교 상세
-	public HighSchoolVO getHighSchoolById(Long hsId);
+	public HighSchoolVO getHighSchoolById(int hsId);
 
 	// 특정 이름으로 고등학교를 조회하는 메서드 추가
 	public List<HighSchoolVO> getHighSchoolsByName(String schoolName);
